Change front end to always display betting options. Users that aren't logged in cannot bet. Users that don't have their emails verified cannot bet. In these scenarios we add text above the button telling them so.
The suggestion box now also always displays but the placeholder text
will change depending on the user being logged in/email verified. They
cannot click the button if they aren't logged in and verified.

Added user status checks on the backend for suggestions and
placing bets, because we were only prohibiting this on the front end.

Light test refactoring